Added Betcoin game definition & install modules from directory rather than db.

+++ src/sql/215.sql
@@ -1,5 +1,5 @@
 ALTER TABLE `events` ADD `season_index` INT NULL DEFAULT NULL AFTER `event_index`;
 ALTER TABLE `events` ADD INDEX (`game_id`, `season_index`);
 ALTER TABLE `game_defined_events` ADD `season_index` INT NULL DEFAULT NULL AFTER `event_index`;
-ALTER TABLE `games` ADD `target_option_block_score` INT NULL DEFAULT NULL AFTER `order_options_by`;
+ALTER TABLE `games` ADD `target_option_block_score` FLOAT NULL DEFAULT NULL AFTER `order_options_by`;
 ALTER TABLE `options` ADD `target_score` FLOAT NULL DEFAULT NULL AFTER `option_block_score`;
